% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APtools.R
\docType{package}
\name{APtools}
\alias{APtools}
\alias{APtools-package}
\title{Functions to estimate Average Positive Predictive Values (AP) for Binary Outcomes and Censored Event.}
\description{
We provide tools to estimate two prediction performance metrics, AP and AUC for risk scores or marker. The outcome of interest is either binary or censored event time. Note that for censored event time, the AP and the AUC are time-dependent for pre-specified time interval(s). A function that compares the APs of two risk scores/markers is also included. Optional outputs include positive predictive values and true positive fractions at the specified marker cut-off values, and a plot of the time-dependent AP versus time (available for event time data).
}

