\name{mayo}
\alias{mayo}
\title{Mayo Marker data}
\description{Two marker values with event time and censoring status for the subjects in Mayo PBC data}

\format{A data frame with 312 observations and 4 variables: time (event time/censoring time), censor (censoring indicator),
mayoscore4, mayoscore5. The two scores are derived from 4 and 5
covariates respectively.}

\source{T Therneau, P Grambsch  (2000)
  Modeling Survival Data: Extending the Cox Model \emph{Springer-Verlag}, New York, ISBN: 0-387-98784-3.
}

\references{Fleming T, Harrington D.  (1991)
  Counting Processes and Survival Analysis \emph{Wiley}, New York.

  Heagerty, P.J., Zheng, Y.  (2005)
  Survival Model Predictive Accuracy and ROC Curves \emph{Biometrics}, \bold{61}, 92 -- 105
}
\keyword{survival}
